backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic-settings==2.7.1
httpx==0.28.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Azure AI Services
azure-ai-documentintelligence==1.0.0
azure-storage-blob==12.24.1
azure-ai-translation-text==1.0.1
azure-cognitiveservices-speech==1.42.0
openai==1.58.1

# Auth
firebase-admin==6.6.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0
